@@ -6,14 +6,14 @@ dependencies:
     sdk: flutter
   audioplayers:
     path: ../
-  http: ^0.12.2
-  path_provider: ^1.5.1
-  path_provider_macos: ^0.0.4+4
-  provider: 4.3.3
+  http: ^0.13.0
+  path_provider: ^2.0.1
+  path_provider_macos: ^2.0.0
+  provider: 5.0.0
 
 dev_dependencies:
-  flutter_driver:
-    sdk: flutter
+  # flutter_driver:
+  #   sdk: flutter
   test: any
 
 flutter:
@@ -25,4 +25,4 @@ flutter:
     - assets/messenger.mp3
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
